Move duplicated code to shared module
