+++ src/app/phase-bug-reporting/phase-bug-reporting-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from '../core/guards/auth.guard';
 import { CanDeactivateIssueGuard } from '../core/guards/can-deactivate-issue-guard.service';
-import { IssueComponent } from './issue/issue.component';
+import { IssueComponent } from '../shared/issue/issue/issue.component';
 import { NewIssueComponent } from './new-issue/new-issue.component';
 import { PhaseBugReportingComponent } from './phase-bug-reporting.component';
