+++ src/app/phase-bug-reporting/phase-bug-reporting.module.ts
@@ -6,24 +6,23 @@ import { IssueComponentsModule } from '../shared/issue/issue-components.module';
 import { LabelDropdownModule } from '../shared/label-dropdown/label-dropdown.module';
 import { SharedModule } from '../shared/shared.module';
 import { ViewIssueModule } from '../shared/view-issue/view-issue.module';
-import { IssueComponent } from './issue/issue.component';
 import { NewIssueComponent } from './new-issue/new-issue.component';
 import { PhaseBugReportingRoutingModule } from './phase-bug-reporting-routing.module';
 import { PhaseBugReportingComponent } from './phase-bug-reporting.component';
-import { IssuesPostedComponent } from './issues-posted/issues-posted.component';
-import { IssuesDeletedComponent } from './issues-deleted/issues-deleted.component';
+import { IssuePageModule } from '../shared/issue-page/issue-page.module';
 
 @NgModule({
   imports: [
     PhaseBugReportingRoutingModule,
     SharedModule,
     IssueComponentsModule,
+    IssuePageModule,
     CommentEditorModule,
     ViewIssueModule,
     MarkdownModule.forChild(),
     IssueTablesModule,
     LabelDropdownModule
   ],
-  declarations: [PhaseBugReportingComponent, NewIssueComponent, IssueComponent, IssuesPostedComponent, IssuesDeletedComponent]
+  declarations: [PhaseBugReportingComponent, NewIssueComponent]
 })
 export class PhaseBugReportingModule {}
