--- src/app/phase-bug-trimming/issue/issue.component.css
@@ -1,16 +1,0 @@
-.row {
-  display: flex;
-}
-
-.column {
-  padding: 10px;
-  height: 100%;
-}
-
-.left {
-  width: 85%;
-}
-
-.right {
-  width: 15%;
-}
