--- src/app/phase-bug-trimming/issue/issue.component.html
@@ -1,3 +1,0 @@
-<div>
-  <app-view-issue [issueId]="this.issueId" [issueComponents]="this.issueComponents"></app-view-issue>
-</div>
