--- src/app/phase-bug-trimming/issues-deleted/issues-deleted.component.css
@@ -1,3 +1,0 @@
-::ng-deep .text-strike-through > mat-table > mat-row {
-  text-decoration: line-through;
-}
