--- src/app/phase-bug-trimming/issues-deleted/issues-deleted.component.html
@@ -1,25 +1,0 @@
-<div data-testid="app-issues-deleted">
-  <mat-grid-list cols="3" rowHeight="80px">
-    <mat-grid-tile>
-      <div class="grid-flush-left">
-        <h1 class="mat-headline" style="margin: 0px">
-          {{ userService.currentUser.role === 'Student' ? 'Issues you deleted' : 'All Deleted Issues' }}
-        </h1>
-      </div>
-    </mat-grid-tile>
-
-    <mat-grid-tile>
-      <mat-form-field class="full-grid-width">
-        <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Search" />
-      </mat-form-field>
-    </mat-grid-tile>
-  </mat-grid-list>
-
-  <app-issue-tables
-    class="text-strike-through"
-    table_name="tableIssuesDeleted"
-    [headers]="this.displayedColumns"
-    [actions]="this.actionButtons"
-    [filters]="this.filter"
-  ></app-issue-tables>
-</div>
