+++ src/app/phase-bug-trimming/phase-bug-trimming.module.ts
@@ -2,14 +2,13 @@ import { NgModule } from '@angular/core';
 import { SharedModule } from '../shared/shared.module';
 import { PhaseBugTrimmingRoutingModule } from './phase-bug-trimming-routing.module';
 import { PhaseBugTrimmingComponent } from './phase-bug-trimming.component';
-import { IssueComponent } from './issue/issue.component';
 import { ViewIssueModule } from '../shared/view-issue/view-issue.module';
-import { IssuesDeletedComponent } from './issues-deleted/issues-deleted.component';
-import { IssuesPostedComponent } from './issues-posted/issues-posted.component';
 import { IssueTablesModule } from '../shared/issue-tables/issue-tables.module';
+import { IssueComponentsModule } from '../shared/issue/issue-components.module';
+import { IssuePageModule } from '../shared/issue-page/issue-page.module';
 
 @NgModule({
-  imports: [PhaseBugTrimmingRoutingModule, SharedModule, ViewIssueModule, IssueTablesModule],
-  declarations: [PhaseBugTrimmingComponent, IssueComponent, IssuesDeletedComponent, IssuesPostedComponent]
+  imports: [PhaseBugTrimmingRoutingModule, SharedModule, ViewIssueModule, IssueTablesModule, IssueComponentsModule, IssuePageModule],
+  declarations: [PhaseBugTrimmingComponent]
 })
 export class PhaseBugTrimmingModule {}
