+++ src/app/shared/issue-page/issue-page.module.ts
@@ -1,0 +1,11 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { IssueComponent } from '../issue/issue/issue.component';
+import { ViewIssueModule } from '../view-issue/view-issue.module';
+
+@NgModule({
+  imports: [CommonModule, ViewIssueModule],
+  declarations: [IssueComponent],
+  exports: [IssueComponent]
+})
+export class IssuePageModule {}
