+++ src/app/shared/issue/issue/issue.component.css
