+++ src/app/shared/issue/issue/issue.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { ISSUE_COMPONENTS, ViewIssueComponent } from '../../shared/view-issue/view-issue.component';
+import { ISSUE_COMPONENTS, ViewIssueComponent } from '../../view-issue/view-issue.component';
 
 @Component({
   selector: 'app-issue',
