+++ src/app/shared/issue/issues-deleted/issues-deleted.component.css
