+++ src/app/shared/issue/issues-posted/issues-posted.component.html
